matval_core/pyproject.toml
[build-system]
requires = ["uv_build>=0.5.15"]
build-backend = "uv_build"

[project]
name = "matval-core"
version = "0.1.0"
description = "Shared database connectivity for Matval"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.build-backend]
source = "src"
